# server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PTCF Room & Equipment Reservation System API",
    "description": "API documentation for the UPLB ICropS Plant Tissue Culture Facility reservation management system.",
    "version": "1.0.0",
    "contact": {
      "name": "PTCF Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-render-url.onrender.com/api",
      "description": "Production server (Render)"
    }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all users (Admin only)",
        "description": "Returns all registered users with their id, email, accountType, userCategory, and createdAt. Requires system_admin role.",
        "operationId": "listUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of user objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "email": {
                        "type": "string",
                        "example": "user@uplb.edu.ph"
                      },
                      "accountType": {
                        "type": "string",
                        "example": "regular_user"
                      },
                      "userCategory": {
                        "type": "string",
                        "example": "student"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/admin/users/{id}/role": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update user role (Admin only)",
        "description": "Promotes or demotes a user's accountType. Admin cannot change their own role.",
        "operationId": "updateUserRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "integer",
              "example": 2
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountType"
                ],
                "properties": {
                  "accountType": {
                    "type": "string",
                    "enum": [
                      "regular_user",
                      "ptcf_staff",
                      "system_admin"
                    ],
                    "example": "ptcf_staff"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User role updated successfully"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "accountType": {
                          "type": "string"
                        },
                        "userCategory": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid role or self-demotion attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/audit-logs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List recent audit logs (Admin only)",
        "description": "Returns recent append-only booking event audit entries. Supports filtering by eventType, bookingId, actorUserId, and limit.",
        "operationId": "listAuditLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum rows to return (1 to 200, default 50)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "required": false,
            "description": "Filter by booking event type",
            "schema": {
              "type": "string",
              "example": "booking.created"
            }
          },
          {
            "name": "bookingId",
            "in": "query",
            "required": false,
            "description": "Filter by booking ID",
            "schema": {
              "type": "integer",
              "example": 123
            }
          },
          {
            "name": "actorUserId",
            "in": "query",
            "required": false,
            "description": "Filter by actor user ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 45
                          },
                          "eventId": {
                            "type": "string",
                            "example": "f5f7a1ee-80b0-4318-91fd-f4f8d3f654bf"
                          },
                          "eventType": {
                            "type": "string",
                            "example": "booking.created"
                          },
                          "occurredAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "topic": {
                            "type": "string",
                            "example": "booking-events"
                          },
                          "partition": {
                            "type": "integer",
                            "example": 0,
                            "nullable": true
                          },
                          "offset": {
                            "type": "string",
                            "example": "98",
                            "nullable": true
                          },
                          "actorUserId": {
                            "type": "integer",
                            "nullable": true
                          },
                          "bookingId": {
                            "type": "integer",
                            "nullable": true
                          },
                          "resourceType": {
                            "type": "string",
                            "enum": [
                              "equipment",
                              "room"
                            ],
                            "nullable": true
                          },
                          "resourceId": {
                            "type": "integer",
                            "nullable": true
                          },
                          "bookingType": {
                            "type": "string",
                            "enum": [
                              "pencil",
                              "firm"
                            ],
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "nullable": true
                          },
                          "payload": {
                            "type": "object"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "actor": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "email": {
                                "type": "string",
                                "format": "email"
                              },
                              "accountType": {
                                "type": "string"
                              }
                            }
                          },
                          "booking": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "resourceType": {
                                "type": "string",
                                "enum": [
                                  "equipment",
                                  "room"
                                ]
                              },
                              "resourceId": {
                                "type": "integer"
                              },
                              "bookingType": {
                                "type": "string",
                                "enum": [
                                  "pencil",
                                  "firm"
                                ]
                              },
                              "status": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/admin/users/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a user account (Admin only)",
        "description": "Permanently deletes a user account. Admin cannot delete their own account.",
        "operationId": "deleteUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "integer",
              "example": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete own account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid credentials"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized - Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingToken": {
                "value": {
                  "message": "Missing Bearer token"
                }
              },
              "invalidToken": {
                "value": {
                  "message": "Invalid or expired token"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden - Insufficient role (requires staff or admin)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Forbidden: insufficient role"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter ONLY your JWT token (without 'Bearer' prefix). Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    }
  },
  "security": []
}